LayoutTests/media/volume-activate-audio-session-expected.txt


EXPECTED (internals.audioSessionActive() == 'false') OK
RUN(video.volume = 0)
RUN(video.src = findMediaFile("audio", "content/silence"))
RUN(video.play())
EXPECTED (internals.audioSessionActive() == 'false') OK
RUN(video.volume = 1)
EXPECTED (internals.audioSessionActive() == 'true') OK
END OF TEST
